SystemUI: fix no service and emergency calls only placement
When the quick settings are not fully expanded and the carrier indicator is displaying no service or emergency calls only, their positions are offset to the left by the width of the signal indicator which isn't there and will not be there for the full pulldown animation either.

This fixes that and a case where if you're connected to a network, press airplane mode in the fully expanded header and then quickly unexpand the header, the carrier name will move left and out of its home where it belongs due to its icon running away.

Change-Id: I55e2d9603a1bb7c47c21eb78d71e55cdc9cd7078
Signed-off-by: Akhil Narang <akhilnarang.1999@gmail.com>