Stop USB Debugging authorization window dismissing on tap outside.
This dialog is onerous to trigger (requires disconnecting and
reconnecting the device), and is too easy to dismiss accidentally
by slightly mis-tapping (happens to me all the time).

This code is mostly abandoned - it has barely been touched at
all since it was introduced in 2012.

This CL makes it not dismiss when the user clicks outside the
dialog. To dismiss, the user now has to click either
CANCEL or ALLOW.

Test: Manually checked that clicking outside the dialog
      dismissed it before but not after this CL.

Change-Id: I603bba9c79e0df8a52ba7db323fea3a13acaa0a5
Signed-off-by: Anirudh Gupta <anirudhgupta109@gmail.com>